- Refactor useStarInitiator.tsx to become a function components instead of a hook.
